Use Study predicates to fetch studies by taxonomy id or name
instead of going though Taxonomy predicates which require dependency on
TaxonomyRepository. Furthermore that would have required an additional
step to fetch from the database after we have the taxonomy and all its
descendents related to the taxonomy id or name for all the studies that
have taxonomies in that immediate list.