Add allowableValues for enum parameter
not ideal, but Swagger does not seem to support adding the enum list via
reference to the enum type definition itself. Here by adding the repeated
enum list as an allowableValues attribute on @ApiImplicitParam the Swagger
UI for the endpoint will have a dropdown list for selection for the parameter.