Add PATCH a study endpoint
Previously because of the aspects for release date restriction and
deprecation, there is no way to undeprecate a deprecated study or update
an existing study's release date if the study's release date is already
set in a date in the future.

This change modifies the @Around advices for StudyRepository.findOne(..)
and update the join point to CrudRepository.findOne(ID), so that we can
use QuerydslPredicateExecutor.findOne(Predicate) to fetch a study object.

Added a new endpoint at /studies/{id}/patch to allow update a study's
properties, regardless of its release date and deprecated status.